+++ lib/anki_record/deck/deck_attributes.rb
@@ -3,26 +3,26 @@
 module AnkiRecord
   # Module with the Deck class's attribute readers, writers, and accessors.
   module DeckAttributes
-    attr_reader :anki21_database
+    attr_reader :anki21_database # :nodoc:
 
     ##
-    # The deck's name.
+    # The deck's name
     attr_accessor :name
 
     ##
-    # The deck's description.
+    # The deck's description
     attr_accessor :description
 
     ##
-    # The deck's id.
+    # The deck's id
     attr_reader :id
 
     ##
-    # The number of seconds since the 1970 epoch when the deck was last modified.
+    # The number of seconds since the 1970 epoch when the deck was last modified
     attr_reader :last_modified_timestamp
 
     ##
-    # The deck's deck options group object.
+    # The deck's deck options group
     attr_reader :deck_options_group
   end
 end