+++ lib/anki_record/note_type/note_type_attributes.rb
@@ -2,60 +2,60 @@
 
 module AnkiRecord
   ##
-  # Module with the NoteType class's attribute readers, writers, and accessors.
+  # Module with the NoteType class's attribute readers, writers, and accessors
   module NoteTypeAttributes
     attr_reader :anki21_database, :latex_svg, :tags, :req, :vers
 
     ##
-    # The note type's id.
+    # The note type's id
     attr_reader :id
 
     ##
-    # The note type's name.
+    # The note type's name
     attr_accessor :name
 
     ##
-    # A boolean that indicates if this note type is a cloze-deletion note type.
+    # A boolean that indicates if this note type is a cloze-deletion note type
     attr_accessor :cloze
 
     ##
-    # The number of seconds since the 1970 epoch at which the note type was last modified.
+    # The number of seconds since the 1970 epoch at which the note type was last modified
     attr_reader :last_modified_timestamp
 
     ##
-    # The note type's update sequence number.
+    # The note type's update sequence number
     attr_reader :usn
 
     ##
-    # The note type's sort field.
+    # The note type's sort field
     attr_reader :sort_field
 
     ##
-    # The note type's CSS.
+    # The note type's CSS
     attr_accessor :css
 
     ##
-    # The note type's LaTeX preamble.
+    # The note type's LaTeX preamble
     attr_reader :latex_preamble
 
     ##
-    # The note type's LaTeX postamble.
+    # The note type's LaTeX postamble
     attr_reader :latex_postamble
 
     ##
-    # The note type's card template objects, as an array.
+    # The note type's card template objects, as an array
     attr_reader :card_templates
 
     ##
-    # The note type's field objects, as an array.
+    # The note type's field objects, as an array
     attr_reader :note_fields
 
     ##
-    # The note type's deck's id.
+    # The note type's deck's id
     attr_reader :deck_id
 
     ##
-    # The note type's deck.
+    # The note type's deck
     def deck
       return nil unless @deck_id
 
@@ -63,7 +63,7 @@ module AnkiRecord
     end
 
     ##
-    # Sets the note type's deck object.
+    # Sets the note type's deck object
     def deck=(deck)
       unless deck.instance_of?(AnkiRecord::Deck)
         raise ArgumentError,